% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_says.R
\name{mark_says}
\alias{mark_says}
\title{Return a Mark quote}
\usage{
mark_says()
}
\value{
A character string containing Mark's quote.
}
\description{
This function prints the well-known quote:
\emph{"So that's unknown, unknown, Delaware, unknown, unknown... That's a perfect score!"}
}
\examples{
mark_says()
}
